Adicionada implementação da classe SensorController. Notificações do sensor via bluetooth finalizada.
Alterado nome visível do dispositivo.

Documentação atualizada.

+++ app/BluetoothController.h
@@ -24,7 +24,7 @@ private:
 	ServerCallbacks *mServerCallbacksP = nullptr;
 	bool mOldDeviceConnected = false;
 	bool mDeviceConnected = false;
-	uint16_t mSensorValue = 0;
+	float mSensorValue = 0.0f;
 	String mId;
 
 	// Variáveis especiais de espera não bloqueante
